Move db tests to unit folder
- Move tests from db folder to up to unit folder
- Remove db folder
- Add lint_test.py to unit folder
- Fix test in index_tests
