+++ tests/unit/fixtures/smile.jpg
